requirements.txt
Flask>=2.3.2
gunicorn>=21.2.0
importlib-metadata
packaging>=24.1
parallel-ssh>=2.12.0
prometheus_client>=0.17.0
pyyaml
pyzmq
setuptools-git-versioning>=2.0,<3
